cerveau_orchestrateur/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
pydantic-settings
httpx
langchain
langgraph
langchain-ollama
langsmith>=0.1.99,<0.2
python-dotenv
pandas
openpyxl
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
google-auth
boto3
scikit-learn
